+++ R/buildtriggers.R
@@ -259,6 +259,17 @@ parse_buildtrigger_list <- function(x) {
 #' )
 #' }
 #'
+#' # create a pubsub trigger that uses github as a source of code to build upon
+#' gh <- cr_buildtrigger_repo("MarkEdmondson1234/googleCloudRunner")
+#' blist <- cr_build_make(cr_build_yaml(cr_buildstep_r('list.files()')))
+#'
+#' \dontrun{
+#' cr_buildtrigger(blist,
+#'                 name = "pubsub-triggered-github-source",
+#'                 trigger = pubsub_trigger,
+#'                 sourceToBuild = gh)
+#' }
+#'
 cr_buildtrigger <- function(build,
                             name,
                             trigger,
@@ -269,6 +280,7 @@ cr_buildtrigger <- function(build,
                             includedFiles = NULL,
                             trigger_tags = NULL,
                             projectId = cr_project_get(),
+                            sourceToBuild = NULL,
                             overwrite = FALSE) {
   assert_that(
     is.string(name),
@@ -293,6 +305,11 @@ cr_buildtrigger <- function(build,
     # the_build <- cr_build_make(build)
   }
 
+  if (!is.null(sourceToBuild)) {
+    assert_that(is.buildtrigger_repo(sourceToBuild))
+    sourceToBuild <- as.gitRepoSource(sourceToBuild)
+  }
+
   trigger_cloudsource <- NULL
   trigger_github <- NULL
   trigger_pubsub <- NULL
@@ -300,6 +317,9 @@ cr_buildtrigger <- function(build,
 
   if (is.gar_pubsubConfig(trigger)) {
     trigger_pubsub <- trigger
+    if(is.null(sourceToBuild)){
+      cli::cli_alert_warning("No sourceToBuild detected for PubSub trigger")
+    }
   } else if (is.buildtrigger_repo(trigger)) {
     # trigger params
     if (trigger$type == "github") {
@@ -307,8 +327,16 @@ cr_buildtrigger <- function(build,
     } else if (trigger$type == "cloud_source") {
       trigger_cloudsource <- trigger$repo
     }
+
+    if(!is.null(sourceToBuild)){
+      stop("Can't use sourceToBuild for git event based triggers, needs to be Webhook or PubSub trigger",
+           call. = FALSE)
+    }
   } else if (is.gar_webhookConfig(trigger)) {
     trigger_webhook <- trigger
+    if(is.null(sourceToBuild)){
+      cli::cli_alert_warning("No sourceToBuild detected for PubSub trigger")
+    }
   } else {
     stop("We should never be here - something wrong with trigger parameter", call. = FALSE)
   }
@@ -326,6 +354,7 @@ cr_buildtrigger <- function(build,
     tags = trigger_tags,
     disabled = disabled,
     substitutions = substitutions,
+    sourceToBuild = sourceToBuild,
     ignoredFiles = ignoredFiles,
     includedFiles = includedFiles
   )
@@ -435,6 +464,7 @@ cr_buildtrigger_copy <- function(buildTrigger,
                                  build = NULL,
                                  ignoredFiles = NULL,
                                  github = NULL,
+                                 sourceToBuild = NULL,
                                  substitutions = NULL,
                                  includedFiles = NULL,
                                  disabled = NULL,
@@ -465,6 +495,8 @@ cr_buildtrigger_copy <- function(buildTrigger,
   }
   if (!is.null(disabled)) buildTrigger$disabled <- disabled
 
+  if(!is.null(sourceToBuild)) buildTrigger$sourceToBuild <- sourceToBuild
+
   buildTrigger <- as.BuildTrigger(buildTrigger)
   url <- sprintf(
     "https://cloudbuild.googleapis.com/v1/projects/%s/triggers",
@@ -477,156 +509,4 @@ cr_buildtrigger_copy <- function(buildTrigger,
   stopifnot(inherits(buildTrigger, "BuildTrigger"))
 
   f(the_body = buildTrigger)
-}
-
-#' BuildTrigger Object
-#'
-#' Configuration for an automated build in response to source repositorychanges.
-#'
-#' @param substitutions A named list of Build macro variables
-#' @param filename Path, from the source root, to a file whose contents is used for the build
-#' @param name User assigned name of the trigger
-#' @param tags Tags for annotation of a `BuildTrigger`
-#' @param build Contents of the build template
-#' @param ignoredFiles ignored_files and included_files are file glob matches extended with support for "**".
-#' @param github a \link{GitHubEventsConfig} object - mutually exclusive with \code{triggerTemplate}
-#' @param includedFiles If any of the files altered in the commit pass the ignored_files
-#' @param disabled If true, the trigger will never result in a build
-#' @param triggerTemplate a \link{RepoSource} object - mutually exclusive with \code{github}
-#' @param description Human-readable description of this trigger
-#' @param pubsubConfig PubsubConfig describes the configuration of a trigger that creates a build whenever a Pub/Sub message is published.
-#' @param webhookConfig WebhookConfig describes the configuration of a trigger that creates a build whenever a webhook is sent to a trigger's webhook URL.
-#'
-#' @seealso \url{https://cloud.google.com/build/docs/api/reference/rest/v1/projects.triggers}
-#'
-#' @return BuildTrigger object
-#'
-#' @family BuildTrigger functions
-#' @export
-BuildTrigger <- function(filename = NULL,
-                         name = NULL,
-                         tags = NULL,
-                         build = NULL,
-                         ignoredFiles = NULL,
-                         github = NULL,
-                         substitutions = NULL,
-                         includedFiles = NULL,
-                         disabled = NULL,
-                         triggerTemplate = NULL,
-                         webhookConfig = NULL,
-                         description = NULL,
-                         pubsubConfig = NULL) {
-  assert_that(
-    xor(is.null(build), is.null(filename)),
-    !is.null(webhookConfig) ||
-      !is.null(pubsubConfig) ||
-      xor(is.null(github), is.null(triggerTemplate))
-  )
-
-  if (!is.null(github)) {
-    assert_that(is.gar_GitHubEventsConfig(github))
-  }
-
-  if (!is.null(triggerTemplate)) {
-    assert_that(is.gar_RepoSource(triggerTemplate))
-  }
-
-  structure(rmNullObs(list(
-    filename = filename,
-    name = name,
-    tags = tags,
-    build = build,
-    ignoredFiles = ignoredFiles,
-    github = github,
-    substitutions = substitutions,
-    includedFiles = includedFiles,
-    disabled = disabled,
-    triggerTemplate = triggerTemplate,
-    pubsubConfig = pubsubConfig,
-    webhookConfig = webhookConfig,
-    description = description
-  )),
-  class = c("BuildTrigger", "list")
-  )
-}
-
-is.gar_BuildTrigger <- function(x) {
-  inherits(x, "BuildTrigger")
-}
-
-as.BuildTrigger <- function(x) {
-  assert_that(is.buildTriggerResponse(x))
-
-  BuildTrigger(
-    filename = x$filename,
-    name = x$name,
-    tags = x$tags,
-    build = x$build,
-    ignoredFiles = x$ignoredFiles,
-    github = x$github,
-    substitutions = x$substitutions,
-    includedFiles = x$includedFiles,
-    disabled = x$disabled,
-    triggerTemplate = x$triggerTemplate,
-    description = x$description
-  )
-}
-
-
-#' Pubsub Config (Build Trigger)
-#'
-#' PubsubConfig describes the configuration of a trigger that creates a build whenever a Pub/Sub message is published.
-#'
-#' @param subscription Output only. Name of the subscription.
-#' @param topic The name of the topic from which this subscription is receiving messages.
-#' @param serviceAccountEmail Service account that will make the push request.
-#' @param state Potential issues with the underlying Pub/Sub subscription configuration. Only populated on get requests.
-#'
-#' @return A PubsubConfig object
-#' @seealso `https://cloud.google.com/build/docs/api/reference/rest/v1/projects.locations.triggers#BuildTrigger.PubsubConfig`
-#'
-#' @export
-PubsubConfig <- function(subscription = NULL,
-                         topic = NULL,
-                         serviceAccountEmail = NULL,
-                         state = NULL) {
-  structure(rmNullObs(
-    list(
-      subscription = subscription,
-      topic = topic,
-      serviceAccountEmail = serviceAccountEmail,
-      state = state
-    )
-  ),
-  class = c("gar_pubsubConfig", "list")
-  )
-}
-
-is.gar_pubsubConfig <- function(x) {
-  inherits(x, "gar_pubsubConfig")
-}
-
-#' WebhookConfig (Build Triggers)
-#'
-#' WebhookConfig describes the configuration of a trigger that creates a build whenever a webhook is sent to a trigger's webhook URL.
-#'
-#' @param state Potential issues with the underlying Pub/Sub subscription configuration. Only populated on get requests.
-#' @param secret Resource name for the secret required as a URL parameter.
-#'
-#' @return A WebhookConfig object
-#'
-#' @export
-WebhookConfig <- function(secret, state = NULL) {
-  structure(rmNullObs(
-    list(
-      secret = secret,
-      state = state
-    )
-  ),
-  class = c("gar_webhookConfig", "list")
-  )
-}
-
-is.gar_webhookConfig <- function(x) {
-  inherits(x, "gar_webhookConfig")
 }